Add postProces method to allow post processing.
This allows plugins that want to take advantage of persistent filter's
performance characteristics, but still need to have some code ran for
initial build persistent cache hits (since `processString` isn't called
again on initial build when cache is warmed).

The motivating use-case is for linters. Users of linting tools generally
expect to see all linting output in the console for every initial build.
Prior to this change, that was not possible.

With this change, it would be possible for a linting tool to embed data
in its emitted output so that it can have enough information to generate
these `console.log` statements by simply reading the cached value
provided to the hook.

This changes the in-cache representation to always be JSON with an
`.output` property containing the string output. This allows or
returning any `JSON.stringy`able object from `processString` and receive
that object in `postProcess` for both cache hits and misses.