getDestFilePath should return null for directories matching extensions
In practice, most directories do not have extensions and as such return null when passed to getDestFilePath in a filter with extensions provided. This is important because canProcessFile by default uses this to determine whether a node should be processed. However, in the case where a directory happens to have a matched extension (e.g. the tree contains loader.js/loader.js and we are doing some processing on js files), we only want the loader.js file, and not the loader.js directory, to be processed.